feat: Add news creation and editing endpoints
