+++ app/news/create/page.tsx
@@ -1,16 +1,5 @@
-import { CreateNewsForm } from '@everynews/components/news/create-news-form'
-import { PageHeader } from '@everynews/components/ui/page-header'
+import { CreateNewsForm } from '@everynews/app/news/create/form'
 
-export default function CreateNewsPage() {
-  return (
-    <div className='container py-8'>
-      <PageHeader
-        title='Create News Item'
-        description='Set up a new news source to track'
-      />
-      <div className='mt-8'>
-        <CreateNewsForm />
-      </div>
-    </div>
-  )
-}
+const CreateNewsPage = () => <CreateNewsForm />
+
+export default CreateNewsPage
