+++ components/ui/page-header.tsx
@@ -6,28 +6,23 @@ import { ReactNode } from 'react'
 
 interface PageHeaderProps {
   title: string
-  description?: string
   className?: string
   actions?: ReactNode
 }
 
-export const PageHeader = ({
-  title,
-  description,
-  className,
-  actions,
-}: PageHeaderProps) => {
+export const PageHeader = ({ title, className, actions }: PageHeaderProps) => {
   return (
     <div
       className={cn(
-        'flex flex-col md:flex-row md:items-center md:justify-between gap-4',
+        'flex flex-col md:flex-row md:items-center md:justify-between gap-4 p-4 border-b',
         className,
       )}
     >
-      <SidebarTrigger />
-      <div className='space-y-1.5'>
-        <h1 className='text-2xl font-bold tracking-tight'>{title}</h1>
-        {description && <p className='text-muted-foreground'>{description}</p>}
+      <div className='flex items-center gap-2'>
+        <SidebarTrigger />
+        <div className='space-y-1.5'>
+          <h1 className='text-2xl font-bold tracking-tight'>{title}</h1>
+        </div>
       </div>
       {actions && <div className='flex items-center gap-2'>{actions}</div>}
     </div>
